pyproject.toml
[tool.poetry]
name = "ai_platform"
version = "0.1.0"
description = "An API "
authors = ["Sergio Soto Núñez <sergio.soto.nunez@gmail.com>"]
readme = "README.md"
package-mode = true
packages = [{include = "ai_platform"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.1"}
diffusers = {extras = ["torch"], version = "^0.30.0"}
torch = "^2.4.0"
torchvision = "^0.19.0"
transformers = "^4.44.1"
pydantic-settings = "^2.4.0"
motor = "^3.5.1"
celery = {extras = ["redis"], version = "^5.4.0"}


[tool.poetry.group.development.dependencies]
accelerate = "^0.33.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ai_platform_api = "ai_platform.api.main:app"
images_creation = "ai_platform.task_queue.images_creation:main"
